pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "discordpy-bot-template-with-commands-extension"
version = "0.1.0"
[tool.poetry.dependencies]
Flask = "^2.2.2"
"discord.py" = "^1.3.4"
python = "^3.8"
replit = "^3.2.5"
[tool.poetry.dev-dependencies]
